requirements.txt.txt
asgiref==3.6.0
astroid==2.14.2
colorama==0.4.6
dill==0.3.6
dj-database-url==1.3.0
Django==4.1.7
django-filter==23.1
django-heroku==0.3.1
djangorestframework==3.14.0
Faker==18.3.1
gunicorn==20.1.0
isort==5.12.0
lazy-object-proxy==1.9.0
Markdown==3.4.1
mccabe==0.7.0
platformdirs==3.0.0
psycopg2==2.9.6
pylint==2.16.2
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typed-ast==1.5.4
typing_extensions==4.5.0
tzdata==2022.7
validate-docbr==1.10.0
whitenoise==6.4.0
wrapt==1.14.1